+++ packages/core/lib/addons.ts
@@ -1311,10 +1311,12 @@ export const stylingSettings = (
       priority: 0,
     },
     {
-      id: 'styles.hover',
+      key: 'styles.hover',
       type: 'group',
       label: (t: GetTextCallback) =>
         t('core.styling.hover.title', 'Hover styles'),
+      condition: (element: ElementObject) =>
+        element?.type === 'button',
       fields: [
         ...stylingSettingsFields('styles.hover'),
         {
@@ -1328,7 +1330,7 @@ export const stylingSettings = (
       ],
     },
     {
-      id: 'styles.active',
+      key: 'styles.active',
       type: 'group',
       label: (t: GetTextCallback) =>
         t('core.styling.active.title', 'Active styles'),