+++ packages/core/lib/index.ts
@@ -2,5 +2,6 @@ export { default as Builder } from './Builder';
 
 export * from './addons';
 export * from './classes';
+export * from './utils';
 
 export type * from './types';