+++ packages/react/lib/Editable/Field.tsx
@@ -1,22 +1,23 @@
-import type {
-  ComponentObject,
-  ComponentSettingsFieldObject,
-  ElementObject,
-  FieldObject,
-  FieldOverride,
-  SettingOverride,
-} from '@oakjs/core';
-import type {
-  SpecialComponentPropsWithoutRef,
-} from '@junipero/react';
 import {
   type MutableRefObject,
   useMemo,
 } from 'react';
+import {
+  type ComponentObject,
+  type ComponentSettingsFieldObject,
+  type ElementObject,
+  type FieldObject,
+  type FieldOverride,
+  type SettingOverride,
+  assignDefined,
+} from '@oakjs/core';
+import {
+  type SpecialComponentPropsWithoutRef,
+  omit,
+} from '@junipero/react';
 
 import type { EditableRef } from './index';
 import { useBuilder } from '../hooks';
-import { assignDefined } from '../utils';
 import DynamicComponent from '../DynamicComponent';
 
 export interface FieldProps extends SpecialComponentPropsWithoutRef {
@@ -57,15 +58,19 @@ const Field = ({
   const setting = useMemo(() => assignDefined<typeof fieldSetting>(
     { type: fieldSetting.type },
     fieldSetting,
-    overrides.settings?.toObject?.() || overrides.settings || {},
-    overrides.field?.toObject?.() || overrides.field || {},
+    omit(overrides.settings?.toObject?.() || overrides.settings || {}, ['key']),
+    omit(overrides.field?.toObject?.() || overrides.field || {}, ['key']),
   ), [fieldSetting, overrides, addons]);
 
   const fieldProps = {
     id: setting.id,
     name: setting.name,
     disabled: setting.disabled,
-    value: builder.getElementSettings(element, setting.key, setting.default),
+    value: builder.getElementSettings(
+      element,
+      fieldSetting.key,
+      setting.default
+    ),
     required: setting.required,
     onChange: (overrides?.field as FieldOverride)?.onChange
       ? onCustomChange.bind(null, setting.key, overrides.field)
@@ -87,7 +92,7 @@ const Field = ({
       }
       { ...fieldProps }
       fieldOptions={{
-        onChange: onChange.bind(null, setting.key),
+        onChange: onChange.bind(null, fieldSetting.key),
         field,
         setting,
         overrides: overrides.field,