+++ packages/react/lib/index.stories.tsx
@@ -506,3 +506,23 @@ export const withMultipleEditableTabs = () => (
     onChange={action('change')}
   />
 );
+
+export const withImageUpdload = () => (
+  <Builder
+    onImageUpload={
+      (e: Event) => {
+        action('upload')(e);
+
+        return {
+          name: 'capture.png',
+          // eslint-disable-next-line max-len
+          url: 'https://storage.googleapis.com/poool-dev-cdn/uploads/57f3c14d0b26241cde521669/original/Capture%20d%E2%80%99e%CC%81cran%202023-03-16%20a%CC%80%2010.32.11.png',
+        };
+      }
+    }
+    addons={[baseAddon()]}
+    value={baseContent}
+    options={{ debug: true }}
+    onChange={action('change')}
+  />
+);