+++ packages/strapi-plugin/CHANGELOG.md
@@ -3,6 +3,38 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.8.5](https://github.com/p3ol/oak/compare/v3.8.4...v3.8.5) (2024-10-09)
+
+**Note:** Version bump only for package @oakjs/strapi-plugin
+
+
+
+
+
+## [3.8.4](https://github.com/p3ol/oak/compare/v3.8.3...v3.8.4) (2024-09-27)
+
+**Note:** Version bump only for package @oakjs/strapi-plugin
+
+
+
+
+
+## [3.8.3](https://github.com/p3ol/oak/compare/v3.8.2...v3.8.3) (2024-09-27)
+
+**Note:** Version bump only for package @oakjs/strapi-plugin
+
+
+
+
+
+## [3.8.2](https://github.com/p3ol/oak/compare/v3.8.1...v3.8.2) (2024-09-27)
+
+**Note:** Version bump only for package @oakjs/strapi-plugin
+
+
+
+
+
 # [3.8.0](https://github.com/p3ol/oak/compare/v3.7.13...v3.8.0) (2024-09-26)